New Framed
refactored code to work with the improved framed struct. The struct uses
std::map. The mappedKeyPoints is now a map that maps the labels to the
keypoints for the label. mappedLabels maps the indices of the keypoints.
roiClusterCount maps the labels of the sample features to the count of
how many sample features have the label.